Render error message on GradebookClient error
