+++ lib/gradebook_client/configuration.rb
@@ -6,20 +6,4 @@ module GradebookClient
       @adapter = Adapters::Fake.new
     end
   end
-
-  def self.configuration
-    @configuration ||= Configuration.new
-  end
-
-  def self.configuration=(config)
-    @configuration = config
-  end
-
-  def self.configure
-    yield configuration
-  end
-
-  def self.adapter
-    configuration.adapter
-  end
 end
